+++ src/components/Record/Record.tsx
@@ -7,7 +7,7 @@ import RecordMenu from '../RecordMenu/RecordMenu'
 import Editable from './Editable'
 import { updateTask, ROOT_ID } from '../../utils/taskService'
 import RecordList from '../RecordList/RecordList'
-import { store, useEvent } from '../../classes/Store2'
+import { store, useEvent } from '../../classes/Store'
 import metadata from '../../classes/Metadata'
 
 type Props = {
